Changed course access and course classes
Modified the course access to the database so that courses can be added
to the database based on a course objected, and modified in a similar way.
Updated the course class to match the functionality required in the
database access. Fixed some errors in AssistantAdmin.java arising
from previous changes.