Further updates to course selection
Modified course selection so that is now a switch case taking
an integer argument. So far the possiblities for creating a course,
modifying a course, deleting a course, and adding an activity to a
course are covered in the switch case, with a 4th option for marking
that is not yet implemented.