Attempted #216, no testing.
Assist #220:  A marking page should be able to be accessed now provided
all the necessary objects exist in the database.

Assist #215: Necessary functionality for marking should be in place so
we can also test createcourse

Assist #217: Marking page should be accessible now based on the function
I created.  No testing though. @mbalaski please look it over quickly,
change it if need be

Assist #187: Should be possible now to navigate to marking page given
the correct 'objects' exist in database.