Updates to course and account functionality
Added some work to Ian's TA box that should make it work (relatively)
well. TA list (should) now update properly based off of the entries in
the box Ian added for multiple TAs. Also changed some functionality of
accounts in relation to the database, so for example if a TA is deleted
all of their entries in the TA table will also be removed. Some of these
changes might not work perfectly, so if any errors arise in account
modification or working with courses, let me know.