Test Suite Skeleton implemented
Also button from MarkingCode page should correctly navigate to the test
suite.

One thing I was not sure about was if the Activity object being passed
into MarkingCode was the specific activity pertaining to that student or
the activity in general.

If the activity is specifically pertaining to the student the ok_button
action event on MarkingCode will have to be updated to put this specific
value into the constructor for TestSuite.  As it stands right now this
constructor just takes the activity object that was passed into
MarkingCode page.