Undid some previous work on account modification
I realised that users are supposed to be able to have duplicate accounts
for different roles, so I undid some of the work I did previously the other
day (deleting or changing type of instructors/TAs). The text fields for
instructor ID and the TA list on the create course page are now uneditable,
and there will have to be a "Delete TA" button on the create course page
so that no invalid data can be entered in the database. This is because
I had to remove the foreign key constraints on the tables (InstructorID
and Name no longer reference accounts, same for TA table) so that we could
have multiple accounts for one person.